Add reset date

+++ src/model.py
@@ -128,7 +128,7 @@ class Simstrat(object):
             shutil.rmtree(self.simulation_dir)
         if not os.path.exists(self.simulation_dir):
             os.makedirs(self.simulation_dir, exist_ok=True)
-        if os.path.exists(os.path.join(self.simulation_dir, "Results")):
+        if os.path.exists(os.path.join(self.simulation_dir, "Results")) and self.args["remove_existing_results"]:
             shutil.rmtree(os.path.join(self.simulation_dir, "Results"))
         os.makedirs(os.path.join(self.simulation_dir, "Results"), exist_ok=True)
         os.chmod(os.path.join(self.simulation_dir, "Results"), 0o777)
@@ -159,6 +159,8 @@ class Simstrat(object):
             self.create_par_file()
             if self.args["run"]:
                 self.run_simulation()
+                if self.args["reset_date"]:
+                    self.reset_date()
                 if self.args["post_process"]:
                     self.post_process()
                     if self.args["upload"]:
@@ -170,7 +172,7 @@ class Simstrat(object):
                 self.log.info("Removing input and output files of failed run (debug=False)")
                 for root, dirs, files in os.walk(self.simulation_dir):
                     for file in files:
-                        if file.endswith(".dat"):
+                        if file.endswith(".dat") or file.endswith(".nml"):
                             os.remove(os.path.join(root, file))
             raise ValueError("Processing failed. See log for details.")
 
@@ -483,6 +485,8 @@ class Simstrat(object):
             self.log.info("Running from {} - {}".format(self.start_date, self.end_date), indent=1)
             run_subprocess(command)
             if os.path.exists(snapshot_path):
+                snapshot_out_path = os.path.join(self.simulation_dir, "simulation-snapshot_{}.dat".format(self.end_date.strftime("%Y%m%d")))
+                shutil.copy(snapshot_path, snapshot_out_path)
                 os.remove(snapshot_path)
         self.log.end_stage()
 
@@ -510,4 +514,13 @@ class Simstrat(object):
         remote_folder = os.path.join(self.args["results_folder_api"], self.key)
         upload_files(local_folder, remote_folder, self.args["server_host"], self.args["server_user"],
                      self.args["server_password"])
+        self.log.end_stage()
+
+    def reset_date(self):
+        self.log.begin_stage("reset_date")
+        if self.snapshot:
+            start_date = self.forcing_start
+            if start_date < self.parameters["reference_date"]:
+                start_date = self.parameters["reference_date"]
+            overwrite_par_file_dates(os.path.join(self.simulation_dir, "Settings.par"), start_date, self.end_date, self.parameters["reference_date"])
         self.log.end_stage()
